Move some common logic to a helpers class
