+++ tracdap-libs/tracdap-lib-orch/src/main/java/org/finos/tracdap/common/cache/CacheTicket.java
@@ -19,9 +19,14 @@ package org.finos.tracdap.common.cache;
 
 import java.time.Duration;
 import java.time.Instant;
+import java.time.temporal.ChronoUnit;
 
 
 public final class CacheTicket implements AutoCloseable {
+
+    public static final int FIRST_REVISION = 0;
+    public static final Duration DEFAULT_DURATION = Duration.of(30, ChronoUnit.SECONDS);
+    public static final Duration MAX_DURATION = Duration.of(5, ChronoUnit.MINUTES);
 
     private static final Instant NO_EXPIRY = Instant.MAX;
